refactor: extracts calendar mapping utils
