+++ projects/client/src/lib/sections/lists/stores/_internal/mapToActivityCalendar.ts
@@ -1,52 +1,7 @@
 import type { Calendar } from '$lib/features/calendar/models/Calendar.ts';
-import { getDayKey } from '$lib/utils/date/getDayKey.ts';
-
-const DAYS_IN_WEEK = 7;
-
-type ActivityEntry<T> = T & ({ watchedAt: Date } | { activityAt: Date });
-
-function getActivityTime<T>(entry: ActivityEntry<T>): Date {
-  if ('watchedAt' in entry) {
-    return entry.watchedAt;
-  }
-
-  return entry.activityAt;
-}
-
-function groupHistoryByDay<T>(items: ActivityEntry<T>[]) {
-  return items.reduce((groups, item) => {
-    const key = getDayKey(getActivityTime(item));
-    const group = groups.get(key) ?? [];
-
-    group.push(item);
-    groups.set(key, group);
-
-    return groups;
-  }, new Map<string, ActivityEntry<T>[]>());
-}
-
-function createHistoryCalendar<T>(
-  groups: Map<string, ActivityEntry<T>[]>,
-  startDate: Date,
-) {
-  if (!startDate) return [];
-
-  return Array.from({ length: DAYS_IN_WEEK })
-    .map((_, i) => {
-      const date = new Date(startDate);
-      date.setDate(date.getDate() + i);
-      const key = getDayKey(date);
-      const items = groups.get(key) ?? [];
-
-      return {
-        date,
-        items: items.toSorted((a, b) =>
-          getActivityTime(b).getTime() - getActivityTime(a).getTime()
-        ),
-      };
-    })
-    .sort((a, b) => a.date.getTime() - b.date.getTime());
-}
+import type { ActivityEntry } from '$lib/sections/lists/stores/models/ActivityEntry.ts';
+import { createHistoryCalendar } from './createHistoryCalendar.ts';
+import { groupHistoryByDay } from './groupHistoryByDay.ts';
 
 export function mapToActivityCalendar<T>(
   items: ActivityEntry<T>[],
