+++ projects/client/src/lib/sections/lists/stores/models/ActivityEntry.ts
@@ -1,0 +1,1 @@
+export type ActivityEntry<T> = T & ({ watchedAt: Date } | { activityAt: Date });
